setup edit product functionality
